Basic functionality for divisions done.
Extended database schema to support information for pool, division, and game type.
This allows for better tracking of when the round robin play is finished and what games
should be included in the standings.

New URL handler added /div/<division> to get information for specific division only.

API extended to accped CALL/<division> to filter on division as well