Major refactor and multitournament support
Major refactor from single large python file to app and module base structure.

Tournaments is now a class which has enabled the ability to display and run multiple tournaments at a time. To facilitate this all the URL paths have changed.

Almost all functionality was maintained. Seeded pods was not brought forward as the previous implimentation was a mess. This will need to be re-implemented at some point