Basic page based user auth
Using Flask-Login all /admin pages are now protected for user login.

Three levels of user now exist
* Regular user - can be assigned as tournament admin from the tournament direction, could also fill other future roles
* Admin user - has permission to make tournament changes on all tournaments (head ref, trusted board members ect)
* Site admin - Highest level, currently treated as admin user, additional seperation for possible future site-wide administration

Additionally, tournaments now have a concept of "active" to a tournament. Unless set to true (1) none of the admin functions for the tournament will be available like updating scores.

Update SQL schema for new users table. Additionally extended tournaments table with new concept of "active" and "admin_ids" which will be a comma separated list of user_ids that are allowed to administrate the tournament.