Fixes to support https/http based on origin
Fixes into view/edit timing rules where it needs to make API calls
to populate fields and needs to know if it should use HTTP or HTTPS.

Can't use requests because production https site is behind proxy and
all calls are http by the time it makes it to the app. Instead, look
for localhost in request.host and act accordingly.

Fixes #46