More app restructure and tests
Restructure the way that the flask app starts to better follow best
practice which allows for using pytest-flask with pytest to test the app
with context. Not sure I've got it all figured out but making progress.

Added test functions to allow for defining a schedule and scores in a
json to load in for testing. Should allow for creating various
tournament conditions and testing standings/ties/rankings/seeding and
all that.