Player search
Added a `/players` page with player search. New `/cgi/search` end point will be used for all search types, so far only players is implemented. This endpoint is subject to change to meet the needs of the site so its not part of `/api`.

Additionally cleaned up some layout and formatting, especially the footer.