Refactoring game notation from silly string to json
Previously who was playing white/black in the database were stored as a 
string that had to be parsed apart via regex to decode team, seed, 
winner/looser notation and it was a mess. Now we use a JSON that defines 
the type and the parameters needed for that type. Much more straight 
forward and avoids issues like numbers in group names getting confused 
with seeed. 

Same notation used for final placings now too. 

Started to remove some places where the partern of returning negativr 
enumbers for errors or unknowns and replacing with None or exceptions.